Decides exclusions based on plaintext paths.
Since exclusions will support wildcards, it's not possible to store
the list of exclusions as a list of encrypted paths (calculated during
initialization). Instead, the plain paths are stored, and the path to
be checked is decrypted at the time of isExcluded check and compared
with the stored plaintext paths.

Function to remove .name suffix from LongNameFilename.

--exclude now supports gitignore-like patterns with wildcards.

Removed duplicated calls to decryptPath.

isExcluded needs to decryptPath, so it was made to return the results
of this function to that the Fuse callbacks don't need to call
decryptPath again.

Move functions related to file exclusion (reverse mode) to new file.

Adds -exclude-wildcard command-line option.

-exclude keeps the original behaviour by matching from the root of the
mounted filesystem.

The new option -exclude-wildcard (or -ew) matches anywhere (unless tha
pattern begins with '/').

Adds -exclude-from command-line option.

It allows the user to specify a file containg exclusion patters. The
patterns are "wildcard" ones, that is, matching anywhere.

Documents new exclusion options.

Optimizes OpenDir.

Files are excluded before their names are encrypted, to avoid having
the names decrypted again when checking for exclusions.

Make NameTransform an interface.

This was done to allow it to be replace with a mock object in tests
(specifically in excluder_test.go).

Exclude DirIV files if parent directory is excluded.

Assuming a directory SOMEDIR (possibly deep in the filesystem
hierarchy) is excluded, now SOMEDIR/gocryptfs.diriv is considered
excluded.

If it were not excluded, it might be possible to infer the existence
of SOMEDIR by checking for SOMEDIR/gocryptfs.diriv.

In practice I've not been able to actually exploit this, as fuse or
the kernel tries to open each directory in a path before actually
trying to open (or stat) a file, but it's safer if gocryptfs makes a
further check.

Fail StatFs if isExcludedCipher returns an error.

Exclude long name .name files if parent directory is excluded.

Adds function to detect file type.

Rename isExcludedCiper to getFileInfo.

getFileInfo returns the file type.

Use file type returned by getFileInfo in GetAttr, Access and Open.

Fail OpenDir and Readlink if target is a virtual file.